Updated xml-adapter-tests
