+++ test/scripts/adapters/xml/async/index.js
@@ -1,0 +1,9 @@
+// jshint esversion: 8
+
+// Test createDatabaseValue
+describe('XmlFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
