+++ test/scripts/adapters/xml/index.js
@@ -1,0 +1,6 @@
+// jshint esversion: 8
+
+describe('xml', () => {
+  require('./async');
+  require('./sync');
+});
