+++ test/scripts/adapters/xml/sync/index.js
@@ -1,0 +1,9 @@
+// jshint esversion: 8
+
+// Test createDatabaseValue
+describe('SyncXmlFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
